Initial Implementation of japanese audio game

+++ AndroidTest/app/src/main/java/red/guih/games/japanese/CompareAnswers.java
@@ -7,22 +7,22 @@ import java.util.stream.Collectors;
 
 public class CompareAnswers {
 
-	public static double compare(String x, String y) {
+	public static float compare(String x, String y) {
 
 		String s = Objects.toString(x, "");
 		String s2 = Objects.toString(y, "");
 
 		List<String> arrayList = bigrams(s);
 		List<String> arrayList2 = bigrams(s2);
-		double nx = arrayList.size();
-		double ny = arrayList2.size();
+		float nx = arrayList.size();
+        float ny = arrayList2.size();
 		arrayList2.removeAll(arrayList);
 
 		if (nx + ny == 0) {
 			return 1;
 		}
 
-		return 1 - 2.0 * arrayList2.size() / (nx + ny);
+		return 1 - 2 * arrayList2.size() / (nx + ny);
 	}
 
 	private static List<String> bigrams(String s) {
